+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [Unreleased]
 
 - Updated `README.md` links given GitHub organisation move
+- Changed keyboard font to [Stroke Input Keyboard v1.0.2]
 
 
 ## [v0.1.0] Implemented drawing (2021-06-27)
@@ -13,6 +14,9 @@ of the drawing part of AOSP's `Keyboard.java` and `KeyboardView.java`.
 Note that the keyboard is literally a bunch of drawn rectangles;
 the actual functionality has not been implemented yet.
 
+
+[Stroke Input Keyboard v1.0.2]:
+  https://github.com/stroke-input/stroke-input-font/releases/tag/v1.0.2
 
 [Unreleased]: https://github.com/stroke-input/stroke-input-android/compare/v0.1.0...HEAD
 [v0.1.0]: https://github.com/stroke-input/stroke-input-android/releases/tag/v0.1.0